Negative Numbers & Unary Operations
Added support for unary operations using a stack (every requested unary operation is added to the stack, then when a number is appended the entire stack is performed on it). Under the hood this adds an unary minus and a square root, but the GUI is the same, so there is a minus button but no square root button for now. Adding buttons is not complicated on a technical level, but requires some design decisions to make.